+++ plugins/admission/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 1.6.0
+* adds `namespaceMetadata` field to `metadata` report 
+
 ## 1.5.6
 * upgrade plugins on build